requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
openmeteo-requests
requests-cache
retry-requests
pandas
python-jose[cryptography]
passlib[bcrypt]
shapely
geopandas
rasterio
scikit-learn
joblib
numpy
matplotlib
seaborn
